% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/assay.R
\name{compute.metrics}
\alias{compute.metrics}
\title{Compute copy-number metrics.}
\usage{
compute.metrics(marker.set.id, params, db, cell.lines = NULL,
  platform = "Illumina", normalize = TRUE)
}
\arguments{
\item{marker.set.id}{ID of marker set containing markers that match params}

\item{params}{reference parameters (see note)}

\item{db}{Database connection.}

\item{cell.lines}{Sample set ID for cell line samples, or NULL to select all cell lines.}

\item{platform}{Platform name.}

\item{normalize}{whether to normalize the intensity values.}
}
\value{
list of metrics (one per sample)
}
\description{
Compute copy-number metrics (BAF and LRR) for cell line samples based on
intensity values.
}
\details{
Computing LRR requires a large number of reference samples. We provide reference parameters
for the MUGA and MegaMUGA array. A future version of the software will provide tools to
derive new reference parameters. In the meantime, please contact the author for help adapting
new arrays to work with this software.
}

