% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/copy_number.R
\name{copy.number.genoCN.plot}
\alias{copy.number.genoCN.plot}
\title{Plot the BAF, LRR and copy number states of a sample.}
\usage{
copy.number.genoCN.plot(metrics, result, cn.range = NULL, ...)
}
\arguments{
\item{metrics}{copy number metrics}

\item{result}{from calling copy.number.genoCN}

\item{cn.range}{the range of copy numbers to display}

\item{...}{additional arguments to pass to \code{plot.metrics}.}
}
\description{
Plot the BAF, LRR and copy number states of a sample.
}

