% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/assay.R
\name{estimate.contamination}
\alias{estimate.contamination}
\title{Estimate contamination using platform-specific (i.e. intensity) information.}
\usage{
estimate.contamination(cl.metrics, dilution.series, baf.thresholds = c(0.02,
  0.46))
}
\arguments{
\item{cl.metrics}{Metrics for cell lines.}

\item{dilution.series}{A DilutionSeries, or a path to an RData file containing a
pre-computed DilutionSeries.}

\item{baf.thresholds}{vector of two BAF thresholds (homozygous and heterozygous).

Estimating BAF thresholds requires a large number of heterozygous control samples.
For our paper, we used the F1 samples in the database to compute the mean mirrored
BAF for homozygous and heterozygous calls and rounded up and down, respectively, to
two decimal places. A future version of this software will provide a more robust
method of determining the thresholds, as well as tools for generating reference
parameters for new arrays. Until then, please contact the author for help adapting
new arrays to work with this software.}
}
\value{
Numeric vector with names as sample IDs and values as predictions as the fraction
of contaminant cells.
}
\description{
Estimate contamination using platform-specific (i.e. intensity) information.
}

