% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/assay.R
\name{execute.assay}
\alias{execute.assay}
\title{Execute an assay on a set of cell lines.}
\usage{
execute.assay(assay, db, error.rate, cell.lines = NULL, impute.bgs = T,
  results.per.cl = 10, secondary.threshold = 0.96)
}
\arguments{
\item{assay}{CLASPAssay object}

\item{db}{database connection}

\item{error.rate}{The error rate to use for PIA calculations.}

\item{cell.lines}{sample set IDs or vector of sample IDs of cell lines to test.
If null, all cell lines in the database will be tested.}

\item{impute.bgs}{boolean or vector of background (e.g. inbred strain) names. If
TRUE, synthetic F1s will be imputed for pairwise combinations of all backgrounds
with their impute field set to TRUE.
Only samples with impute column set to TRUE are used. If there is more than one
sample of a given background, one is selected randomly.}

\item{results.per.cl}{Number of matches to report for each cell line.}

\item{secondary.threshold}{match score below which a secondary background is searched.}
}
\value{
The return value is a list with the following elements:
1. samples: Data frame of cell line sample data
2. pairwise.matrix: Pairwise similarity matrix between all pairs of cell lines
3. table: a summary data frame of the results with the following columns:
  id: Cell line sample ID
  primary.match: Name of best matching sample
  primary.match.score: Alignment score of primary match
  secondary.match: Name of best secondary match (if primary.match.score < secondary.threshold)
  secondary.match.score: Alignment score of secondary match
  pia: Probabilty of Incorrect Assignment
4. details: a list the same length as the number of cell lines with the following elements:
  primary.matches
  secondary.matches
}
\description{
Execute an assay on a set of cell lines.
}

