% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.all.arrays}
\alias{get.all.arrays}
\title{Get information for all genotyping arrays in the database.}
\usage{
get.all.arrays(db)
}
\arguments{
\item{db}{the database connection.}
}
\value{
table of array information.
}
\description{
Get information for all genotyping arrays in the database.
}

