% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.all.genomes}
\alias{get.all.genomes}
\title{Get information for all genome builds in the database.}
\usage{
get.all.genomes(db)
}
\arguments{
\item{db}{The database connection.}
}
\value{
table of genome information.
}
\description{
Get information for all genome builds in the database.
}

