% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.array.id}
\alias{get.array.id}
\title{Get array ID by name.}
\usage{
get.array.id(name, db)
}
\arguments{
\item{name}{the array name.}

\item{db}{database connection}
}
\value{
the ID.
}
\description{
Get array ID by name.
}

