% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.genome.id}
\alias{get.genome.id}
\title{Get genome ID by name.}
\usage{
get.genome.id(name, db)
}
\arguments{
\item{name}{genome name}

\item{db}{database connection}
}
\value{
the ID
}
\description{
Get genome ID by name.
}

