% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.genotypes}
\alias{get.genotypes}
\title{Fetch genotype data for arbitrary sets of markers and samples.}
\usage{
get.genotypes(markers, db, samples = NULL, as.matrix = TRUE)
}
\arguments{
\item{markers}{A marker set ID or a vector of marker IDs.}

\item{db}{The database connection.}

\item{samples}{A sample set ID, a list of sample IDs, or NULL for all samples.}

\item{as.matrix}{Whether the result should be converted from a three-column data frame
(marker ID, sample ID, call) to a MxN matrix.}
}
\value{
A MxN data frame of (recoded) genotypes, where M = markers and N = samples. Row names
are marker IDs and colnames are sample IDs.
}
\description{
Fetch genotype data for arbitrary sets of markers and samples.
}

