% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.platform.data}
\alias{get.platform.data}
\title{Fetch platform-specific data.}
\usage{
get.platform.data(marker.set.id, platform, db, samples = NULL,
  as.matrix = TRUE, format = 1)
}
\arguments{
\item{marker.set.id}{The assay ID.}

\item{platform}{The name of the platform-specific data table.}

\item{db}{The database connection.}

\item{samples}{The set of samples, or all samples if NULL.}

\item{as.matrix}{Whether the result should be converted from a four column data frame to a list
of two MxN matricies.}

\item{format}{If as.matrix=TRUE, controls whether rows are samples and columns are markers
(format=1) or vice-versa (format=2).}
}
\value{
A list of data frames. Each is a MxN data frame of a platform-specific variable,
where M = markers and N = samples.
}
\description{
Fetch platform-specific data.
}

