% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{get.samples}
\alias{get.samples}
\title{Get a list of the samples in the set with the specified ID, or by a
vector of sample IDs.}
\usage{
get.samples(ids, db, where = NULL, include.pedigree = TRUE)
}
\arguments{
\item{ids}{The sample set ID, or a list of sample IDs.}

\item{db}{the datbase connection.}

\item{where}{list of key/value pairs to specify conditions for sample
table columns.}

\item{include.pedigree}{whether to return pedigree information in the
sample table.}
}
\value{
data frame with sample information.
}
\description{
Get a list of the samples in the set with the specified ID, or by a
vector of sample IDs.
}

