% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{import.genotypes}
\alias{import.genotypes}
\title{Import genotype data.}
\usage{
import.genotypes(genotypes, db, N.call = "N", platform = NULL,
  marker.set.id = NULL, sample.set.id = NULL)
}
\arguments{
\item{genotypes}{Genotype data. Either a path to a tab-delimited file (supports gzipped files)
or a data.frame. See Details.}

\item{db}{The database connection.}

\item{N.call}{The genotype used for missing information (no-call). Any other genotype that is
not the two possible alleles or the N.call value is considered a heterozygous call.}

\item{platform}{The name of the table in which to store platform-specific data, if applicable.}

\item{marker.set.id}{ID of a marker set. If NULL, marker IDs will be selected by name.}

\item{sample.set.id}{ID of a sample set. If NULL, sample IDs will be selected by name.}
}
\description{
\code{genotypes} must be a data.frame with at least three columns. The first two columns specify
the sample and marker. Each may be either numeric, in which case it is interpreted as an ID, or
character, in which case it is interpreted as a name. The third column is the genotype call.
Subsequent columns are interpreted as platform-specific data.
}
\details{
Genotypes are .recoded into a bit vector (stored as an integer), where each bit represents one of
the possible alleles. Missing information is coded as 0.
}

