% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/database.R
\name{insert.assay}
\alias{insert.assay}
\title{Insert assay into database.}
\usage{
insert.assay(assay, db)
}
\arguments{
\item{assay}{CLASPAssay object.}

\item{db}{database connection.}
}
\value{
the assay updated with an ID column.
}
\description{
Insert assay into database.
}

