% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/viz.R
\name{metrics.plot}
\alias{metrics.plot}
\title{Plot BAF and LRR values for a single sample.}
\usage{
metrics.plot(name, sample.id, marker.set.id, metrics, norm.params, db,
  chromosomes = 1:19, zoom = NULL, tick.interval = NULL,
  plot.baf = TRUE, plot.lrr = TRUE, lrr.ylim = c(-2, 2),
  plot.trend = TRUE, density.band.window = 100, density.band.frac = 0.75,
  other.plots = NULL, other.data = NULL, ...)
}
\arguments{
\item{name}{sample name.}

\item{sample.id}{sampleID.}

\item{marker.set.id}{marker set ID.}

\item{metrics}{data.frame with metrics (BAF and LRR).}

\item{norm.params}{normalization params.}

\item{db}{database connection.}

\item{chromosomes}{the chromosomes to plot.}

\item{zoom}{if plotting a single chromosome, the region to zoom in on.}

\item{tick.interval}{distance between x-axis ticks (in Mb). Defaults to
a single tick per chromosome.}

\item{plot.baf}{whether to plot B allele frequencies.}

\item{plot.lrr}{whether to plot Log R Ratios.}

\item{lrr.ylim}{y-axis range within which to plot LRRs.}

\item{plot.trend}{whether to plot a trend line for LRR values.}

\item{density.band.window}{if non-null, overlays a plot of the region that contains
the specified fraction of values (\code{density.band.frac}) within windows the
specified size.}

\item{density.band.frac}{the fraction of values to include in the density band.}

\item{other.plots}{plotting function(s) to add additional panels to the BAF and LRR plots.}

\item{other.data}{additional data required by the additional plotting functions.}

\item{...}{additional arguments to pass to panel plotting functions.}
}
\description{
\code{data} must be a data.frame with one row per marker and
with at least two columns: chrm and pos (marker chromosome and position).
If, \code{plot.baf=TRUE}, there must be a \code{baf} column, and if
\code{plot.lrr=TRUE}, there must be a \code{lrr} column. If there is
also a \code{call} column, points in the BAF plot will be colored according
to genotype call.
}

