% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/assay.R
\name{summary.CLASPAssay}
\alias{summary.CLASPAssay}
\title{Generate a summary of a CLASPAssay object.}
\usage{
\method{summary}{CLASPAssay}(object, ...)
}
\arguments{
\item{object}{CLASPAssay object.}

\item{...}{unused.}
}
\value{
summary.CLASPAssay object.
}
\description{
Generate a summary of a CLASPAssay object.
}

